Rewrites function names to match PowerShell naming
It involves renaming the function names according to Powershell's
list of Approved Verbs to use in naming actions.
https://learn.microsoft.com/en-us/powershell/scripting/developer/cmdlet/approved-verbs-for-windows-powershell-commands

It also uses the prompt option within the Read-Host command to display
the text to the user before they enter their input.
This reduces the number of lines to write and makes the code more
readable.

Finally, some basic variable name changes to match what is being stored
in it.