Fixes a small issue with checking the tag name
From the looks of it, the function was being called recursively within
the loop.
This resulted in showing errors of a tag name already existing, even
when the tag name was not existing in the list.

So instead of calling the function all over again, we just prompt the
user again and reassign the variable there itself within the loop.