Adds the `Invoke-Main` function
Thus completing the PowerShell script.
There are some general issues with the script, but it is functional.

The script does not properly check the flags for the script.
When I run it, it does not check the flags and just runs the script.
I have to close the program and run it again in order for it to check
the flags.
So it is function in that sense.

Need to look up more on a solution for this.
Will be added as an issue on GitHub too.